# index-server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Dev-connector API",
    "description": "API description format for different endpoints"
  },
  "schemes": ["http", "https"],
  "host": "localhost:80",
  "paths": {
...
    "/api/users/register": {
      "post":
      {
        "summary": "Register User",
        "tags": ["Users"],
        "description": "Route for registering an user to the database",
        "parameters": [
          {
            "name": "RegUser",
            "in": "body",
            "description": "Register credentials",
            "schema": {
              "type": "object",
              "required": ["name", "email", "password", "password2"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "password2": {
                  "type": "string"
                }
              }
            }
          }],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "password2": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "The user has been successfully registered",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "format": "object"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "avatar": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "__v": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "summary": "Login User",
        "tags": ["Users"],
        "description": "Route to logging an user",
        "parameters": [{
          "name": "LogUser",
          "in": "body",
          "description": "Login credentials",
          "schema": {
            "required": ["email", "password"],
            "properties": {
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "The user has successfully logged in",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "The user can't be found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "summary": "Get posts",
        "tags": ["Posts"],
        "description": "Get all posts in descending",
        "responses": {
          "200": {
            "description": "All posts are returned",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Post"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Post",
        "tags": ["Posts"],
        "description": "Create private post by a certain user",
        "parameters": [{
          "name": "PostInfo",
          "in": "body",
          "description": "Initial post parameters info",
          "schema": {
            "type": "object",
            "required": ["text", "name"],
            "properties": {
              "text": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "avatar": {
                "type": "string"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "The post who is created",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        }
      }
    },
    "/api/posts/{postID}": {
      "get": {
        "summary": "Get Post",
        "tags": ["Posts"],
        "description": "Route for getting post by post ID",
        "parameters": [{
          "$ref": "#/parameters/postID"
        }],
        "responses": {
          "200": {
            "description": "The post was found in the posts collection",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "404": {
            "description": "The post with that ID cannot be found",
            "schema": {
              "type": "object",
              "required": ["error"],
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Post",
        "tags": ["Posts"],
        "description": "Deletion of a post by it's id",
        "parameters": [{
          "$ref": "#/parameters/postID"
        }],
        "responses": {
          "200": {
            "description": "The post was successfully deleted",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized to delete someone elses post",
            "schema": {
              "type": "object",
              "properties": {
                "notauthorized": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/like/{postID}": {
      "post": {
        "summary": "Like Post",
        "tags": ["Posts"],
        "description": "Like post by post id",
        "parameters": [{
          "$ref": "#/parameters/postID"
        }],
        "responses": {
          "200": {
            "description": "User likes post and gets added to the array of likes by his ID",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "400": {
            "description": "User already liked this post",
            "schema": {
              "type": "object",
              "properties": {
                "alreadyliked": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "No post found",
            "schema": {
              "type": "object",
              "properties": {
                "postnotfound": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/unlike/{postID}": {
      "post": {
        "summary": "Unlike post",
        "tags": ["Posts"],
        "description": "Unlike post with post ID",
        "parameters": [{
          "$ref": "#/parameters/postID"
        }],
        "responses": {
          "500": {
            "description": "Post was removed",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "400": {
            "description": "You have not liked the post yet",
            "schema": {
              "type": "object",
              "properties": {
                "notliked": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "The post wasn't found",
            "schema": {
              "type": "object",
              "properties": {
                "postnotfound": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "JWT": {
      "description": "JSON WEB TOKEN",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [{
    "JWT": []
  }],
  "definitions": {
...
    "Post": {
      "type": "object",
      "required": ["user", "_id", "text", "name", "date"],
      "properties": {
        "_id": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "user": {
          "type": "string"
        },
        "likes": {
          "type": "array",
          "items": {
            "properties": {
              "user": {
                "type": "string",
                "format": "object"
              },
              "_id": {
                "type": "string",
                "format": "object"
              }
            }
          }
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "date": {
          "type": "string",
          "format": "date"
        }

      }
    },
    "Comment": {
      "required": ["user", "text"],
      "type": "object",
      "properties": {
        "user": {
          "type": "string",
          "format": "object"
        },
        "text": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "date": {
          "type": "string",
          "format": "date"
        }
      }
    },
...
  },
  "parameters": {
    "postID": {
      "name": "postID",
      "in": "path",
      "required": true,
      "description": "The id of the post we are looking for",
      "type": "string"
    },
...
  }
}